+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed `!important` not applying to `outline` https://github.com/Textualize/textual/issues/2420
 - Fixed `outline-right` not being recognised https://github.com/Textualize/textual/issues/2446
 
+### Added
+
+- Watch methods can now optionally be private https://github.com/Textualize/textual/issues/2382
+
 ## [0.22.3] - 2023-04-29
 
 ### Fixed